Login

+++ frontend/src/app/users/user-list/user-list.component.ts
@@ -2,6 +2,7 @@ import {Component, OnInit, Input} from '@angular/core';
 
 import {User} from '../../../models/user.model';
 import {UserService} from '../../../services/user.service';
+import {Router} from "@angular/router";
 
 @Component({
   selector: 'app-user-list',
@@ -11,13 +12,17 @@ import {UserService} from '../../../services/user.service';
 export class UserListComponent implements OnInit {
 
   public userList: User[] = [];
+  private userConnected: User;
 
   @Input()
   displayDeleteButton: boolean;
 
-  constructor(private userService: UserService) {
+  constructor(private router: Router, private userService: UserService) {
     this.userService.users$.subscribe((users: User[]) => {
-      this.userList = users;
+      this.userService.userSelected$.subscribe(user => {
+        this.userConnected = user;
+        this.userList = users;
+      });
     });
   }
 
@@ -26,7 +31,13 @@ export class UserListComponent implements OnInit {
   }
 
   deleteUser(user: User): void {
-    this.userService.deleteUser(user);
+    if (user.id === this.userConnected.id){
+      document.getElementById('error_msg_' + user.id).innerText = 'Vous ne pouvez pas supprimer un utilisateur lorsque vous êtes connécté avec !';
+      setTimeout(() => document.getElementById('error_msg_' + user.id).innerText = '', 2000);
+    }
+    else {
+      this.userService.deleteUser(user);
+    }
   }
 
   selectUser(user: User): void {
@@ -39,5 +50,13 @@ export class UserListComponent implements OnInit {
 
   isAdminFalse(user: User): void {
     this.userService.updateUser(user);
+  }
+
+  createUser(user: User): void {
+    user.login = this.userConnected.login;
+    user.password = this.userConnected.password;
+    this.userService.addUser(user).subscribe(u => {
+      this.userService.addToUsers(u);
+    });
   }
 }
